requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-genai
python-dotenv==1.0.0
streamlit==1.29.0
requests==2.31.0
google-genai
flask
flask-cors
python-dotenv
requests
streamlit
flask
flask-cors
google-genai
python-dotenv
requests
streamlit
flask
flask-cors
python-dotenv
requests
streamlit
